requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-api-python-client==2.147.0
google-auth==2.35.0
firebase-admin==6.5.0
pydantic-settings==2.5.2
apscheduler==3.10.4
pytz==2024.2
